refactor: move execution methods to HoleFiller and FimPromptBuilder
- Move getFromChat to HoleFiller class
- Move getFromFIM to FimPromptBuilder class
- Remove execution methods from GhostInlineCompletionProvider
- Provider now delegates to HoleFiller/FimPromptBuilder for execution
- Pass processSuggestion as callback to maintain encapsulation
- All tests pass (63/63)
